requirements.txt
django
openpyxl
pylint
psycopg2
mysqlclient
django-phonenumber-field
phonenumberslite